updated the list of professional exam

+++ src/app/page/aws-practise/page.tsx
@@ -1,10 +1,17 @@
 'use client';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { aws_question_bank } from '@/data/aws_question';
 import Exam from '@/app/components/questions/Exam';
 import Practise from '@/app/components/practise/Practice';
 const page = () => {
-	localStorage.clear();
+	    useEffect(() => {
+				// Check if localStorage is available
+				if (typeof localStorage === 'undefined') {
+					console.warn('localStorage is not available in this environment.');
+				} else {
+					localStorage.clear();
+				}
+			}, []);
 	return (
 		<div>
 			<Practise
