+++ src/app/page/itil-practise/page.tsx
@@ -1,9 +1,16 @@
 'use client';
 import Practise from '@/app/components/practise/Practice';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { itilQuestions } from '@/data/questionBank';
 const page = () => {
-	localStorage.clear();
+	useEffect(() => {
+		// Check if localStorage is available
+		if (typeof localStorage === 'undefined') {
+			console.warn('localStorage is not available in this environment.');
+		} else {
+			localStorage.clear();
+		}
+	}, []);
 	return (
 		<div>
 			<Practise
